requirements.txt
astroid==2.2.5
aniso8601==6.0.0
Click==7.0
Flask==1.0.3
Flask-Cors==3.0.7
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pytz==2019.1
six==1.12.0
Werkzeug==0.15.5
python-dateutil==2.6.0
pycryptodome== 3.9.9
pylint==2.3.1
six==1.12.0
typed-ast==1.4.2
wrapt==1.11.1
python-jose==3.2.0
astroid==2.2.5
Click==7.0
ecdsa==0.13.2
authlib>=1.0
python-dotenv>=0.19.2
requests>=2.27.1
mccabe==0.6.1
lazy-object-proxy==1.4.0
future==0.17.1
Flask-Script==2.0.6
Flask-Migrate==2.7.0
psycopg2-binary==2.9.1